requirements.txt
fastapi
uvicorn
pydantic
jinja2
python-multipart
aiohttp
requests
python-decouple